Refactor :: enum package location change
