+++ src/main/java/com/keepgoing/keepserver/domain/book/domain/entity/Book.java
@@ -1,7 +1,7 @@
 package com.keepgoing.keepserver.domain.book.domain.entity;
 
 
-import com.keepgoing.keepserver.domain.book.domain.entity.enums.BookState;
+import com.keepgoing.keepserver.domain.book.domain.enums.BookState;
 import com.keepgoing.keepserver.domain.user.domain.entity.user.User;
 import jakarta.persistence.*;
 import lombok.*;
