+++ src/main/java/com/keepgoing/keepserver/domain/book/domain/enums/BookState.java
@@ -1,4 +1,4 @@
-package com.keepgoing.keepserver.domain.book.domain.entity.enums;
+package com.keepgoing.keepserver.domain.book.domain.enums;
 
 public enum BookState {
     AVAILABLE,
