+++ src/main/java/com/keepgoing/keepserver/domain/book/payload/response/BookResponseDto.java
@@ -1,6 +1,6 @@
 package com.keepgoing.keepserver.domain.book.payload.response;
 
-import com.keepgoing.keepserver.domain.book.domain.entity.enums.BookState;
+import com.keepgoing.keepserver.domain.book.domain.enums.BookState;
 import com.querydsl.core.annotations.QueryProjection;
 import lombok.Builder;
