+++ src/main/java/com/keepgoing/keepserver/domain/rent/service/RentServiceImpl.java
@@ -1,11 +1,11 @@
 package com.keepgoing.keepserver.domain.rent.service;
 
 import com.keepgoing.keepserver.domain.book.domain.entity.Book;
-import com.keepgoing.keepserver.domain.book.domain.entity.enums.BookState;
+import com.keepgoing.keepserver.domain.book.domain.enums.BookState;
 import com.keepgoing.keepserver.domain.book.domain.repository.BookRepository;
 import com.keepgoing.keepserver.domain.book.mapper.BookMapper;
 import com.keepgoing.keepserver.domain.device.domain.entity.Device;
-import com.keepgoing.keepserver.domain.device.domain.entity.enums.DeviceStatus;
+import com.keepgoing.keepserver.domain.device.domain.enums.DeviceStatus;
 import com.keepgoing.keepserver.domain.device.domain.repository.DeviceRepository;
 import com.keepgoing.keepserver.domain.device.mapper.DeviceMapper;
 import com.keepgoing.keepserver.domain.device.service.DeviceServiceImpl;
